test fix name collision
